# IGNITE-24134 Sql. Improve performance of KV plan in SQL