--- a/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/KeyValueModifyPlan.java
+++ b/modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/KeyValueModifyPlan.java
@@ -17,6 +17,8 @@
 
 package org.apache.ignite.internal.sql.engine.prepare;
 
+import static org.apache.ignite.internal.sql.engine.util.Commons.cast;
+
 import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -53,6 +55,8 @@ public class KeyValueModifyPlan implements ExplainablePlan, ExecutablePlan {
     private final IgniteKeyValueModify modifyNode;
     private final ResultSetMetadata meta;
     private final ParameterMetadata parameterMetadata;
+
+    private volatile InsertExecution<?> operation;
 
     KeyValueModifyPlan(
             PlanId id,
@@ -112,13 +116,13 @@ public class KeyValueModifyPlan implements ExplainablePlan, ExecutablePlan {
         return modifyNode;
     }
 
-    @Override
-    public <RowT> AsyncCursor<InternalSqlRow> execute(
-            ExecutionContext<RowT> ctx,
-            InternalTransaction tx,
-            ExecutableTableRegistry tableRegistry,
-            @Nullable QueryPrefetchCallback firstPageReadyCallback
-    ) {
+    private <RowT> InsertExecution<RowT> operation(ExecutionContext<RowT> ctx, ExecutableTableRegistry tableRegistry) {
+        InsertExecution<RowT> operation = cast(this.operation);
+
+        if (operation != null) {
+            return operation;
+        }
+
         IgniteTable sqlTable = table();
         ExecutableTable execTable = tableRegistry.getTable(catalogVersion, sqlTable.id());
 
@@ -127,11 +131,25 @@ public class KeyValueModifyPlan implements ExplainablePlan, ExecutablePlan {
         SqlRowProvider<RowT> rowSupplier = ctx.expressionFactory()
                 .rowSource(expressions);
 
-        UpdatableTable updatableTable = execTable.updatableTable();
+        UpdatableTable table = execTable.updatableTable();
 
-        CompletableFuture<Iterator<InternalSqlRow>> result = updatableTable.insert(
-                tx, ctx, rowSupplier.get(ctx)
-        ).thenApply(none -> List.<InternalSqlRow>of(new InternalSqlRowSingleLong(1L)).iterator());
+        operation = new InsertExecution<>(table, rowSupplier);
+
+        this.operation = operation;
+
+        return operation;
+    }
+
+    @Override
+    public <RowT> AsyncCursor<InternalSqlRow> execute(
+            ExecutionContext<RowT> ctx,
+            InternalTransaction tx,
+            ExecutableTableRegistry tableRegistry,
+            @Nullable QueryPrefetchCallback firstPageReadyCallback
+    ) {
+        InsertExecution<RowT> operation = operation(ctx, tableRegistry);
+
+        CompletableFuture<Iterator<InternalSqlRow>> result = operation.perform(ctx, tx);
 
         if (firstPageReadyCallback != null) {
             result.whenComplete((res, err) -> firstPageReadyCallback.onPrefetchComplete(err));
@@ -140,6 +158,24 @@ public class KeyValueModifyPlan implements ExplainablePlan, ExecutablePlan {
         ctx.scheduleTimeout(result);
 
         return new AsyncWrapper<>(result, Runnable::run);
+    }
+
+    private static class InsertExecution<RowT> {
+        private final UpdatableTable table;
+        private final SqlRowProvider<RowT> rowSupplier;
+
+        private InsertExecution(
+                UpdatableTable table,
+                SqlRowProvider<RowT> rowSupplier
+        ) {
+            this.table = table;
+            this.rowSupplier = rowSupplier;
+        }
+
+        CompletableFuture<Iterator<InternalSqlRow>> perform(ExecutionContext<RowT> ctx, InternalTransaction tx) {
+            return table.insert(tx, ctx, rowSupplier.get(ctx))
+                    .thenApply(none -> List.<InternalSqlRow>of(new InternalSqlRowSingleLong(1L)).iterator());
+        }
     }
 
     public int catalogVersion() {