--- a/modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/util/QueryCheckerTest.java
+++ b/modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/util/QueryCheckerTest.java
@@ -67,6 +67,12 @@ public class QueryCheckerTest extends BaseIgniteAbstractTest {
             .build();
 
     @BeforeAll
+    @AfterAll
+    public static void resetFlag() {
+        Commons.resetFastQueryOptimizationFlag();
+    }
+
+    @BeforeAll
     static void startCluster() {
         CLUSTER.start();